Implement search for Content Tags
ContentTag pagination sadly doesn't conform to Zendesk's standards.
It uses cursor pagination, but doesn't include a `links.next` node in the response (which would normally hold the URL of the next page of results).
Because of this, we have to build the 'next page URL' ourselves by extracting the `meta.after_cursor` node value & using it to add a `&page[after]=<cursorValue>` parameter to the original query URL